requirements_app.txt
Flask
#psycopg2
# wheel package will be renamed to below
psycopg2-binary
Flask-WTF
